# web-ui/backend/openapi.yaml
openapi: 3.0.3
info:
  title: AI PowerShell Assistant Web API
  description: RESTful API for AI PowerShell Assistant Web UI
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/history:
    get:
      tags:
        - History
      summary: Get command history
      description: Retrieves paginated command history with optional filtering
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [success, failed, all]
      responses:
        '200':
          description: History retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryListResponse'
  /api/history/{id}:
    get:
      tags:
        - History
      summary: Get history item details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: History item retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HistoryItem'
        '404':
          description: History item not found

    delete:
      tags:
        - History
      summary: Delete history item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: History item deleted
        '404':
          description: History item not found
...
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: User login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              required:
                - username
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      token:
                        type: string
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: User logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
components:
  schemas:
...
    HistoryItem:
      type: object
      properties:
        id:
          type: string
        userInput:
          type: string
        command:
          type: string
        success:
          type: boolean
        output:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        executionTime:
          type: number
        timestamp:
          type: string
          format: date-time
    HistoryListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/HistoryItem'
            total:
              type: integer
            page:
              type: integer
            limit:
              type: integer
...
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT